backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.1",
    "black>=25.1.0",
    "emails>=0.6",
    "fastapi[standard]>=0.115.8",
    "jinja2>=3.1.5",
    "passlib>=1.7.4",
    "psycopg[binary]>=3.2.4",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.10.1",
    "sentry-sdk>=2.21.0",
    "sqlmodel>=0.0.22",
    "tenacity>=9.0.0",
]
